+++ libs/pages/settings/src/lib/ui/page-organization-github-repository-access/page-organization-github-repository-access.spec.tsx
@@ -1,5 +1,4 @@
-import { act, getByTestId, getByText, render } from '__tests__/utils/setup-jest'
-import { GitProviderEnum } from 'qovery-typescript-axios'
+import { renderWithProviders } from '@qovery/shared/util-tests'
 import PageOrganizationGithubRepositoryAccess, {
   type PageOrganizationGithubRepositoryAccessProps,
 } from './page-organization-github-repository-access'
@@ -21,60 +20,7 @@ const props: PageOrganizationGithubRepositoryAccessProps = {
 
 describe('PageOrganizationGithubRepositoryAccess', () => {
   it('should render successfully', () => {
-    const { baseElement } = render(<PageOrganizationGithubRepositoryAccess {...props} />)
+    const { baseElement } = renderWithProviders(<PageOrganizationGithubRepositoryAccess {...props} />)
     expect(baseElement).toBeTruthy()
-  })
-
-  it('should display install button if githubAuthProvider does not use_bot', async () => {
-    const testProps = {
-      ...props,
-      githubAuthProvider: {
-        ...props.githubAuthProvider,
-        use_bot: false,
-      },
-    }
-    const { baseElement } = render(<PageOrganizationGithubRepositoryAccess {...testProps} />)
-
-    await act(() => {
-      getByTestId(baseElement, 'install-button').click()
-    })
-
-    expect(props.onConfigure).toHaveBeenCalled()
-  })
-
-  it('should display disconnect and manage permissions buttons if githubAuthProvider use_bot', async () => {
-    const { baseElement } = render(<PageOrganizationGithubRepositoryAccess {...props} />)
-    const disconnectButton = getByTestId(baseElement, 'disconnect-button')
-    const permissionButton = getByTestId(baseElement, 'permission-button')
-
-    await act(() => {
-      disconnectButton.click()
-      permissionButton.click()
-    })
-
-    expect(props.onDisconnect).toHaveBeenCalled()
-    expect(props.onConfigure).toHaveBeenCalled()
-  })
-
-  it('should display repository list if repositories are not empty and not loading', () => {
-    const testProps: PageOrganizationGithubRepositoryAccessProps = {
-      ...props,
-      repositories: [
-        {
-          id: 'id',
-          name: 'name',
-          url: 'url',
-          provider: GitProviderEnum.GITHUB,
-          is_private: false,
-          default_branch: 'default_branch',
-          branches: {
-            loadingStatus: 'loaded',
-            items: [],
-          },
-        },
-      ],
-    }
-    const { baseElement } = render(<PageOrganizationGithubRepositoryAccess {...testProps} />)
-    getByText(baseElement, 'name')
   })
 })
